Changes in Buses,user,Owner file and add Capacity, CommonServices file.
In Buses file add the Prize to get a ticket of the bus.
In CommonServices File Provide the common services which the owner and
user both using and they also have some Specifice services.
DataBaseConnection file we provide a getConnection method to save the
details of the database and get connection to the same database again
without using the URL , password , username etc